+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Cast resource namespace as literal string for `imports` method.
 - Increase minimum supported version to 1.0.
 - Always return outputs as String for `output` method.
+- Fix `chdir` argument position in `graph` method.
 
 ### 2.1.1
 **Packer**